% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABCMCMCFunctions.R
\name{ABCMCMC}
\alias{ABCMCMC}
\title{Performs and Approximate Bayesian Computation Sampling of Model Parameters}
\usage{
ABCMCMC(objectiveFunction, startPar, nSims, Sigma0, delta, dprior)
}
\arguments{
\item{startPar}{starting value for the parameter vector}

\item{nSims}{requested sample size}

\item{Sigma0}{multivariate normal covariance of Markov chain
transition kernel}

\item{delta}{ABC acceptance threshold}

\item{dprior}{a function that returns prior probability density
values}

\item{experiments}{a list of experiments}

\item{modelName}{model functions will be assumed to have this
prefix (comment(modelName) can contain a file-name)}

\item{parMap}{re-mapping function between the ABC MCMC variables and model
parameters}

\item{getScore}{a function(model_output,experiment) that returns
scores}

\item{nCores}{setting for multicore package}
}
\value{
a list containing a sample matrix and a vector of scores (values of delta for each sample)
}
\description{
Given a set of simulation experiments (list), a model, parameter
boundaries, this function will draw a sample of parameters from the
posterior probability density of the given problem.
}
\details{
Initially this function performs a similar job as an optimizer, and
then transitions to MCMC sampling.
}
