% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{copulaPrior}
\alias{copulaPrior}
\title{copulaPrior creates a prior probability density function}
\usage{
copulaPrior(Copula)
}
\arguments{
\item{Copula}{a list, as returned by fitCopula() or makeIndepCopula}
}
\value{
a function that maps parameters (a vector) to probability density values (scalar)
}
\description{
This function accepts the return list of fitCopula() or
makeIndepCopula() and creates a density function from it.
}
\examples{
prior_pdf<-copulaPrior(makeIndepCopula(ll=c(-1,-1,-1),ul=c(1,1,1)))
prior_pdf(runif(3))
0.125
}
