% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABCMCMCFunctions.R
\name{dprior}
\alias{dprior}
\title{Prior Probability Density Value from Copula Fit}
\usage{
dprior(inx, U, Z, Y, copula, ll, ul)
}
\arguments{
\item{inx}{PDF argument (value of random variable)}

\item{U}{samples of marginal distributions}

\item{Z}{(CDE) cumulative density estimates of U}

\item{Y}{(PDE) probability density estimates of U}

\item{copula}{as returned by copula estimator}

\item{ll}{lower limit of inx}

\item{ul}{upper limit of inx}
}
\value{
probability density function value
}
\description{
This function calculates a probability density value of an argument value x (vector) given a copula
obtained from a sample from the prior (from preceding data sets).
}
\details{
Returns PDF(x)
}
