% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copulaFunctions.R
\name{makeIndepCopula}
\alias{makeIndepCopula}
\title{Copula Formulation for Uniform Prior Distributions}
\usage{
makeIndepCopula(ll, ul)
}
\arguments{
\item{ll}{ll[i] is the lower limit of random variable x[i]}

\item{ul}{upper limit, analogous to ll.}
}
\value{
list with: copula, U, Z, and Y entries.
}
\description{
Covers the (simpler) special case where the prior(x) is iid uniform.
The return value has the same structure as the value of fitCopula().
}
