% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runModel.R
\name{makeObjective}
\alias{makeObjective}
\title{creates Objective functions from ingredients}
\usage{
makeObjective(
  experiments,
  modelName,
  getScore,
  parMap = identity,
  mc.cores = detectCores()
)
}
\arguments{
\item{experiments}{a list of simulation experiments}

\item{modelName}{and model storage file as comment}

\item{getScore}{a function that calculates ABC scores}

\item{parMap}{a function that transforms ABC variables into acceptable model parameters}

\item{mc.cores}{number of cores to use}
}
\value{
an objective function
}
\description{
the returned objective function has only one argument (the ABC
parameters)
}
