% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABCMCMCFunctions.R
\name{checkFitWithPreviousExperiments}
\alias{checkFitWithPreviousExperiments}
\title{ABC acceptance of currently sampled values given old data (Prior)}
\usage{
checkFitWithPreviousExperiments(
  modelName,
  draws,
  experiments,
  parMap = identity(),
  getScore,
  delta,
  nCores = detectCores()
)
}
\arguments{
\item{modelName}{name (prefix), is used to find the file and model functions therein.}

\item{draws}{matrix of sampled values (to be filtered).}

\item{experiments}{a list of experiments (all of them, or up to currentExpSet).}

\item{getScore}{scoring function.}

\item{delta}{the acceptance threshold.}

\item{nCores}{number of cores to use in mclapply() calls.}

\item{currentExpSet}{an index, all experiments before that index
will be evaluated for scores and acceptance.}

\item{experimentsIndices}{can be used to filter simulation
experiments (to exclude some), e.g. 2:4, to exclude 1.}

\item{parVal}{not used other than for size?}

\item{parIdx}{remapping index passed to runModel().}

\item{environment}{passed to runModel(), selects solver.}

\item{nChains}{number of parallel Markov chains (unused?).}
}
\description{
The prior probability density model using copulas and vines is not
perfect, so values sampled from an imperfect prior estimate can be
checked against old data.
}
