% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABCMCMCFunctions.R
\name{parUpdate}
\alias{parUpdate}
\title{Updates Parameter Values}
\usage{
parUpdate(
  experiments,
  modelName,
  parMap,
  curPar,
  canPar,
  curDelta,
  curPrior,
  delta,
  U,
  Z,
  Y,
  copula,
  ll,
  ul,
  getScore,
  environment,
  nCores = detectCores()
)
}
\arguments{
\item{experiments}{list of simulation experiments}

\item{modelName}{(characters), this will be used to find the model
file and the functions in that file}

\item{curPar}{current parameter values (as ABC samples them)}

\item{canPar}{candidate parameter values (for MCMC)}

\item{curDelta}{current distance between data and simulation, if
the MCMC chain has not yet reached any point where this is
below the threshold (delta), this can be accepted as the new
current state for the chain.}

\item{curPrior}{current Prior values given curPar}

\item{delta}{distance threshold for ABC}

\item{U}{sampled marginal parameter values}

\item{Z}{(CDE) cumulative probability estimate of U}

\item{Y}{(PDE) probability density estimate values of U}

\item{copula}{as returned by copula estimators}

\item{ll}{lower limit of parameters}

\item{ul}{upper limit of parameters}

\item{getScore}{a scoring function}

\item{environment}{"C" selects GSL solvers, "R" selects the deSolve
as backend}

\item{nCores}{number of cores to use in mclapply().}

\item{parIdx}{remapping index set for passing parameters to the
model}

\item{parDefVal}{default values for the parameters (for cases in
which some never change)}
}
\value{
updated values for curPar, curDelta, and curPrior
}
\description{
under valid ABC update conditions (successful simulation) the
parameters are updated to new values.
}
