% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABCMCMCFunctions.R
\name{parUpdate}
\alias{parUpdate}
\title{Updates Parameter Values}
\usage{
parUpdate(objectiveFunction, curPar, canPar, curDelta, curPrior, delta, dprior)
}
\arguments{
\item{curPar}{current parameter values (as ABC samples them)}

\item{canPar}{candidate parameter values (for MCMC)}

\item{curDelta}{current distance between data and simulation, if
the MCMC chain has not yet reached any point where this is
below the threshold (delta), this can be accepted as the new
current state for the chain.}

\item{curPrior}{current Prior values given curPar}

\item{delta}{distance threshold for ABC}

\item{dprior}{prior probability density function}

\item{experiments}{list of simulation experiments}

\item{modelName}{(characters), this will be used to find the model
file and the functions in that file}

\item{parMap}{optional remapping function set for passing parameters to the
model: parModel<-parMap(parABC)}

\item{getScore}{a scoring function}

\item{nCores}{number of cores to use in mclapply().}
}
\value{
updated values for curPar, curDelta, and curPrior
}
\description{
under valid ABC update conditions (successful simulation) the
parameters are updated to new values.
}
